+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Changelog
 ## 2020-10-07
+### Changed
+- Adapting the output to the terminal width. 
+
+### Removed
+- Removed `columns` argument.
+
+## 2020-10-06
+### Changed
 - Updated to Swift 5 syntax.
 
 ## 2016-11-13
+### Added
 - Initial version.